+++ device/general/assign-groups-by-template.ps1
@@ -62,7 +62,7 @@ param(
 
 Write-RjRbLog -Message "Caller: '$CallerName'" -Verbose
 
-$Version = "1.0.0"
+$Version = "1.0.1"
 Write-RjRbLog -Message "Version: $Version" -Verbose
 
 Write-RjRbLog -Message "DeviceId: $DeviceId" -Verbose